+++ extension/js/common/api/email-provider/gmail/google.ts
@@ -2,14 +2,13 @@
 
 'use strict';
 
-import { Dict, Str } from '../../../core/common.js';
 import { Api, ProgressCbs, ReqMethod } from '../../shared/api.js';
+import { Dict, Str } from '../../../core/common.js';
 
 import { GMAIL_GOOGLE_API_HOST, PEOPLE_GOOGLE_API_HOST } from '../../../core/const.js';
-import { Serializable } from '../../../platform/store/abstract-store.js';
 import { GmailRes } from './gmail-parser.js';
-import { GoogleAuthHelper } from './google-auth-helper.js';
 import { GoogleAuth } from './google-auth.js';
+import { Serializable } from '../../../platform/store/abstract-store.js';
 
 export class Google {
   public static webmailUrl = (acctEmail: string) => {
@@ -39,7 +38,7 @@ export class Google {
     }
     contentType = contentType || 'application/json; charset=UTF-8';
     // eslint-disable-next-line @typescript-eslint/naming-convention
-    const headers = { Authorization: await GoogleAuthHelper.googleApiAuthHeader(acctEmail) };
+    const headers = { Authorization: await GoogleAuth.googleApiAuthHeader(acctEmail) };
     const xhr = Api.getAjaxProgressXhrFactory(progress);
     const request = { xhr, url, method, data, headers, crossDomain: true, contentType, async: true };
     return (await GoogleAuth.apiGoogleCallRetryAuthErrorOneTime(acctEmail, request)) as RT;
@@ -54,7 +53,7 @@ export class Google {
     const data = { query, readMask: 'names,emailAddresses', pageSize: max };
     const xhr = Api.getAjaxProgressXhrFactory(progress);
     // eslint-disable-next-line @typescript-eslint/naming-convention
-    const headers = { Authorization: await GoogleAuthHelper.googleApiAuthHeader(acctEmail) };
+    const headers = { Authorization: await GoogleAuth.googleApiAuthHeader(acctEmail) };
     const contacts = await Promise.all([
       GoogleAuth.apiGoogleCallRetryAuthErrorOneTime(acctEmail, {
         xhr,
