+++ extension/js/common/api/shared/api.ts
@@ -10,7 +10,7 @@ import { Dict, EmailParts } from '../../core/common.js';
 import { Env } from '../../browser/env.js';
 import { secureRandomBytes } from '../../platform/util.js';
 import { ApiErr, AjaxErr } from './api-error.js';
-import { GoogleAuthHelper } from '../email-provider/gmail/google-auth-helper.js';
+import { GoogleAuth } from '../email-provider/gmail/google-auth.js';
 
 export type ReqFmt = 'JSON' | 'FORM' | 'TEXT';
 export type RecipientType = 'to' | 'cc' | 'bcc';
@@ -214,11 +214,11 @@ export class Api {
       if (ApiErr.isAuthErr(firstAttemptErr) && idToken) {
         // force refresh token
         // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        const { email } = GoogleAuthHelper.parseIdToken(idToken);
+        const { email } = GoogleAuth.parseIdToken(idToken);
         if (email) {
           // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-          req.headers!.Authorization = await GoogleAuthHelper.googleApiAuthHeader(email, true);
-          return (await Api.ajax(req, Catch.stackTrace())) as RT;
+          req.headers!.Authorization = await GoogleAuth.googleApiAuthHeader(email, true);
+          return await Api.ajax(req, Catch.stackTrace()) as RT;
         }
       }
       throw firstAttemptErr;
