+++ extension/js/common/platform/store/acct-store.ts
@@ -1,6 +1,5 @@
 /* ©️ 2016 - present FlowCrypt a.s. Limitations apply. Contact human@flowcrypt.com */
 
-import { GoogleAuthHelper } from '../../api/email-provider/gmail/google-auth-helper.js';
 import { GoogleAuth } from '../../api/email-provider/gmail/google-auth.js';
 import { ApiErr } from '../../api/shared/api-error.js';
 import { BgNotReadyErr, BrowserMsg } from '../../browser/browser-msg.js';
@@ -15,7 +14,7 @@ import { AbstractStore, RawStore } from './abstract-store.js';
 import { InMemoryStore } from './in-memory-store.js';
 
 export type EmailProvider = 'gmail';
-type GoogleAuthScopesNames = [keyof typeof GoogleAuthHelper.OAUTH.scopes, keyof typeof GoogleAuthHelper.OAUTH.legacy_scopes][number];
+type GoogleAuthScopesNames = [keyof typeof GoogleAuth.OAUTH.scopes, keyof typeof GoogleAuth.OAUTH.legacy_scopes][number];
 
 export type Scopes = {
   openid: boolean;
@@ -167,12 +166,12 @@ export class AcctStore extends AbstractStore {
         BrowserMsg.send.notificationShowAuthPopupNeeded('broadcast', { acctEmail });
       }
     }
-    for (const key of Object.keys({ ...GoogleAuthHelper.OAUTH.scopes, ...GoogleAuthHelper.OAUTH.legacy_scopes })) {
+    for (const key of Object.keys({ ...GoogleAuth.OAUTH.scopes, ...GoogleAuth.OAUTH.legacy_scopes })) {
       const scopeName = key as GoogleAuthScopesNames;
-      if (scopeName in GoogleAuthHelper.OAUTH.scopes) {
-        result[scopeName] = allowedScopes.includes(GoogleAuthHelper.OAUTH.scopes[scopeName as keyof typeof GoogleAuthHelper.OAUTH.scopes]);
-      } else if (scopeName in GoogleAuthHelper.OAUTH.legacy_scopes) {
-        result[scopeName] = allowedScopes.includes(GoogleAuthHelper.OAUTH.legacy_scopes[scopeName as keyof typeof GoogleAuthHelper.OAUTH.legacy_scopes]);
+      if (scopeName in GoogleAuth.OAUTH.scopes) {
+        result[scopeName] = allowedScopes.includes(GoogleAuth.OAUTH.scopes[scopeName as keyof typeof GoogleAuth.OAUTH.scopes]);
+      } else if (scopeName in GoogleAuth.OAUTH.legacy_scopes) {
+        result[scopeName] = allowedScopes.includes(GoogleAuth.OAUTH.legacy_scopes[scopeName as keyof typeof GoogleAuth.OAUTH.legacy_scopes]);
       }
     }
     return result;
